+++ Components/IconTypes/IconType_default/default.lua
@@ -72,7 +72,7 @@ function Type:DragReceived(icon, t, data, subType, param4)
 				return
 			end
 			
-			local baseSpellName = GetSpellName(spellData.actionId)
+			local baseSpellName = GetSpellName(spellData.actionID)
 			input = baseSpellName or spellData.name
 		end